Make range array schema reusable
